Removed APIKEY.py from repository but kept locally
Started work on web page that a user can upload their
image to.

In theory the image should be written to local directory
(as of now the app is designed to be used locally by a single user)
in the future I aim to add user_id's and user databases.

The image is then passed to the Preprocess_image module in which it
should write the processed image to local directory /Processed_Files
however there are issues with the file writing, instead I pass the image
from the pre process method to the image reading module.

BasicImageRead should then return the text to the user.
TODO: the text returned from BasicImageRead should be passed
to my API module which deals with finding the important information
and subsequently finding the pokemon card