Processed_Files now returns an array of images.
This array is then used in basic image read until a match is found
images in the array are ordered from most processed to least.

Set up the user input for the set.

Set up regex which searches for 3 digits on either side of a '/'

The numbers and user input will be used to call to an api