Changed layout of code and added saving images to be able to showcase what processing techniques are done on images.
Added api calling and how the recognised text is
read. Without the set name it searches the set
with the same number of total cards and the id of
the current card. With the set name just the id of
the card is used. Thus more likely to get a match.

Call to the api after finding a match and currently
displays image of the card, name of the pokemon and
prices of the different types of cards.

TODO: Database?
Batch processing?
More integration with api calls.